pyproject.toml
[project]
name = "dn410-server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask[async]>=3.1.0",
    "gunicorn>=23.0.0",
    "python-dotenv>=1.0.1",
    "bootstrap-flask>=2.4.1",
    "beautifulsoup4>=4.12.3",
    "asonic>=2.0.0",
    "setuptools",
    "flask-wtf>=1.2.2",
]

[tool.uv.sources]
asonic = { git = "https://github.com/Danie-1/asonic.git", branch = "patch-1" }
